backend/requirements.txt
-i https://pypi.org/simple
alembic==1.14.0; python_version >= '3.8'
apispec[marshmallow]==6.7.1; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
click==8.1.7; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.0
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-migrate==4.0.7; python_version >= '3.6'
flask-smorest==0.45.0; python_version >= '3.9'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.6; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.23.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pyjwt==2.10.0; python_version >= '3.9'
python-dotenv==1.0.1; python_version >= '3.8'
sqlalchemy==2.0.36; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
webargs==8.6.0; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
pymysql==1.0.2
